requirements.txt
alembic==1.0.10
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.1.4
django-celery==3.3.1
django-redis==4.10.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
dwebsocket==0.5.10
fake-useragent==0.1.11
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.19
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==3.0.37
lxml==4.4.0
Mako==1.0.10
MarkupSafe==1.1.1
more-itertools==7.2.0
parsel==1.5.1
Pillow==6.1.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodome==3.9.0
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.9.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
pywin32==224
queuelib==1.5.0
redis==3.3.6
requests==2.22.0
Scrapy==1.7.3
scrapy-djangoitem==1.1.1
scrapy-redis==0.6.8
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.3
sqlparse==0.3.0
Twisted==19.2.1
urllib3==1.25.3
vine==1.3.0
virtualenv==16.7.2
virtualenvwrapper-win==1.2.5
w3lib==1.20.0
Werkzeug==0.15.4
zipp==0.6.0
zope.interface==4.6.0
